Refactor scene blocks
